feat: refactor dashboard menu structure and update routing for improved navigation
